InitializationScript/requirements.txt
Flask>=0.12
azure-cosmos>=1.0.0
cassandra-driver>=1.0.0